requirements.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
Pillow==10.4.0